seo_infrastructure/requirements.txt
openai==1.12.0
requests==2.31.0
spacy==3.7.2
python-dotenv==1.0.0
